+++ src/test/resources/com/atomist/rug/runtime/js/interop/SimpleCommandHandlerReturningMessage.ts
@@ -1,11 +1,11 @@
-import {HandleCommand, Instruction, Response, HandlerContext, Plan, Message} from '@atomist/rug/operations/Handlers'
+import {HandleCommand, Instruction, Response, HandlerContext, Plan} from '@atomist/rug/operations/Handlers'
 import {CommandHandler, Parameter, Tags, Intent} from '@atomist/rug/operations/Decorators'
 
 @CommandHandler("ShowMeTheKitties","Search Youtube for kitty videos and post results to slack")
 class KittieFetcher implements HandleCommand{
 
   handle(ctx: HandlerContext) : Plan {
-    return Plan.ofMessage(new Message("Up and at 'em!"));
+    return Plan.ofMessage({body: "Up and at 'em!", kind: "directed", contentType: "text/plain");
   }
 }