+++ src/test/resources/com/atomist/rug/runtime/js/interop/SimpleCommandHandlerWithNullDefault.ts
@@ -1,4 +1,4 @@
-import {HandleCommand, Instruction, Response, HandlerContext, Plan, Message} from '@atomist/rug/operations/Handlers'
+import {HandleCommand, Instruction, Response, HandlerContext, Plan, ResponseMessage} from '@atomist/rug/operations/Handlers'
 import {CommandHandler, Parameter, Tags, Intent} from '@atomist/rug/operations/Decorators'
 
 @CommandHandler("ShowMeTheKitties","Search Youtube for kitty videos and post results to slack")
@@ -8,7 +8,7 @@ class KittieFetcher implements HandleCommand{
   test: string = null
 
   handle(ctx: HandlerContext) : Plan {
-    return new Plan().add(new Message());
+    return new Plan().add(new ResponseMessage("body"));
   }
 }